Added barriers and started interaction
 - Player can no longer move off the map due to BarrierTiles
 - Now detections for if the player is next to a TileEntity

+++ src/entities/BurgerStation.java
@@ -8,7 +8,7 @@ import tiles.Tile;
 
 public class BurgerStation extends TileEntity{
 
-	public BurgerStation(Handler handler, double x, double y, double z) {
+	public BurgerStation(Handler handler, int x, int y, int z) {
 		super(handler, x, y, z, Assets.burgerStation, Tile.TILEWIDTH * 2, Tile.TILEHEIGHT, Tile.TILEHEIGHT, 2, 1, 1);
 	}
