+++ src/entities/DrinkStation.java
@@ -8,7 +8,7 @@ import tiles.Tile;
 
 public class DrinkStation extends TileEntity{
 
-	public DrinkStation(Handler handler, double x, double y, double z) {
+	public DrinkStation(Handler handler, int x, int y, int z) {
 		super(handler, x, y, z, Assets.drinkStation, Tile.TILEWIDTH, Tile.TILEHEIGHT, Tile.TILEHEIGHT, 1, 1, 2);
 	}
