+++ src/entities/Foodbar.java
@@ -10,7 +10,7 @@ import tiles.Tile;
 public class Foodbar extends TileEntity{
 	
 
-	public Foodbar(Handler handler, double x, double y, double z) {
+	public Foodbar(Handler handler, int x, int y, int z) {
 		super(handler, x, y, z, Assets.foodbar, Tile.TILEWIDTH * 2, Tile.TILEHEIGHT, Tile.TILEHEIGHT, 2, 1, 1);
 		bounds.height = Tile.TILEHEIGHT;
 	}
