+++ src/entities/MilkshakeStation.java
@@ -8,7 +8,7 @@ import tiles.Tile;
 
 public class MilkshakeStation extends TileEntity{
 
-	public MilkshakeStation(Handler handler, double x, double y, double z) {
+	public MilkshakeStation(Handler handler, int x, int y, int z) {
 		super(handler, x, y, z, Assets.milkshakeStation, Tile.TILEWIDTH, Tile.TILEHEIGHT * 2, Tile.TILEHEIGHT, 1, 2, 2);
 	}
